feat(issues): Split out issue event navigation
Fixes an issue where preloading the next event didn't have the correct query parameters, meaning preloading wasn't working.
Splits out the hooks and components for all that preloading stuff into its own component.
