+++ static/app/views/issueDetails/groupDetails.tsx
@@ -238,7 +238,6 @@ function useFetchGroupDetails(): FetchGroupDetailsState {
   } = useGroupEvent({
     groupId,
     eventId: params.eventId,
-    environments,
   });
 
   const {
